# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Crazy Resident A2A Control Plane",
    "version": "0.2.0-dev"
  },
  "paths": {
...
    "/api/runs": {
      "post": {
        "summary": "Create Run",
        "operationId": "create_run_api_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunCreated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/drain": {
      "post": {
        "summary": "Drain Run",
        "operationId": "drain_run_api_runs__run_id__drain_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DrainResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/snapshot": {
      "get": {
        "summary": "Snapshot",
        "operationId": "snapshot_api_snapshot_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/events": {
      "get": {
        "summary": "Events",
        "operationId": "events_api_events_get",
        "parameters": [
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "After"
            }
          },
          {
            "name": "run_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 1,
              "default": 1000,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/events/stream": {
      "get": {
        "summary": "Stream Events",
        "operationId": "stream_events_api_events_stream_get",
        "parameters": [
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "After"
            }
          },
          {
            "name": "run_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          },
          {
            "name": "once",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Once"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chaos/faults": {
      "post": {
        "summary": "Arm Fault",
        "operationId": "arm_fault_api_chaos_faults_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FaultRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FaultResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chaos/peer-probe": {
      "post": {
        "summary": "Peer Probe",
        "operationId": "peer_probe_api_chaos_peer_probe_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PeerProbeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KernelDecision"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentView": {
        "properties": {
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Capabilities"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "mailbox_pending": {
            "type": "integer",
            "title": "Mailbox Pending",
            "default": 0
          },
          "active_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Run Id"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "agent_id",
          "role",
          "status"
        ],
        "title": "AgentView"
      },
      "AssignmentView": {
        "properties": {
          "assignment_id": {
            "type": "string",
            "title": "Assignment Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "goal": {
            "type": "string",
            "title": "Goal"
          },
          "exit_criteria": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Exit Criteria"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "assignment_id",
          "run_id",
          "task_id",
          "agent_id",
          "goal",
          "status"
        ],
        "title": "AssignmentView"
      },
      "CapabilityManifestView": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "turn_id": {
            "type": "string",
            "title": "Turn Id"
          },
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "assignment_id": {
            "type": "string",
            "title": "Assignment Id"
          },
          "strategy": {
            "type": "string",
            "title": "Strategy"
          },
          "catalog_size": {
            "type": "integer",
            "title": "Catalog Size",
            "default": 0
          },
          "disclosed_count": {
            "type": "integer",
            "title": "Disclosed Count",
            "default": 0
          },
          "withheld_count": {
            "type": "integer",
            "title": "Withheld Count",
            "default": 0
          },
          "excluded_count": {
            "type": "integer",
            "title": "Excluded Count",
            "default": 0
          },
          "manifest": {
            "additionalProperties": true,
            "type": "object",
            "title": "Manifest"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "run_id",
          "task_id",
          "turn_id",
          "agent_id",
          "assignment_id",
          "strategy",
          "manifest"
        ],
        "title": "CapabilityManifestView"
      },
      "ContextView": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "trigger_event_id": {
            "type": "string",
            "title": "Trigger Event Id"
          },
          "context_epoch": {
            "type": "integer",
            "title": "Context Epoch"
          },
          "manifest": {
            "additionalProperties": true,
            "type": "object",
            "title": "Manifest"
          },
          "microcompact": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Microcompact"
          },
          "message_preview": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "type": "array",
            "title": "Message Preview"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "run_id",
          "task_id",
          "agent_id",
          "trigger_event_id",
          "context_epoch",
          "manifest",
          "microcompact"
        ],
        "title": "ContextView"
      },
      "DrainResult": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "steps": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Steps"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "steps"
        ],
        "title": "DrainResult"
      },
      "DreamJobView": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "signal_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signal Id"
          },
          "memory_candidate_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Memory Candidate Id"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "job_id",
          "run_id",
          "task_id",
          "status"
        ],
        "title": "DreamJobView"
      },
...
      "EventPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EventRecord"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "EventPage"
      },
...
      "EvolutionView": {
        "properties": {
          "candidate_id": {
            "type": "string",
            "title": "Candidate Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "base_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Version"
          },
          "proposed_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposed Version"
          },
          "rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rationale"
          },
          "evidence_refs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Refs"
          },
          "diffs": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Diffs"
          },
          "metrics": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Metrics"
          },
          "next_gate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Gate"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "candidate_id",
          "run_id",
          "task_id",
          "status"
        ],
        "title": "EvolutionView"
      },
      "FaultRequest": {
        "properties": {
          "point": {
            "type": "string",
            "title": "Point"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "point"
        ],
        "title": "FaultRequest"
      },
      "FaultResult": {
        "properties": {
          "armed": {
            "type": "string",
            "title": "Armed"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          }
        },
        "type": "object",
        "required": [
          "armed",
          "mode"
        ],
        "title": "FaultResult"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KernelDecision": {
        "properties": {
          "candidate_id": {
            "type": "string",
            "title": "Candidate Id"
          },
          "accepted": {
            "type": "boolean",
            "title": "Accepted"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "event_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Event Ids"
          },
          "reused": {
            "type": "boolean",
            "title": "Reused",
            "default": false
          },
          "recovered": {
            "type": "boolean",
            "title": "Recovered",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "candidate_id",
          "accepted",
          "reason"
        ],
        "title": "KernelDecision"
      },
      "MemoryView": {
        "properties": {
          "candidate_id": {
            "type": "string",
            "title": "Candidate Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "slot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slot"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "scope": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope"
          },
          "evidence_refs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Refs"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "risk": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk"
          },
          "admission_zone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Admission Zone"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "candidate_id",
          "run_id",
          "task_id",
          "status"
        ],
        "title": "MemoryView"
      },
      "PeerProbeRequest": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "sender": {
            "type": "string",
            "title": "Sender"
          },
          "receiver": {
            "type": "string",
            "title": "Receiver"
          },
          "depth": {
            "type": "integer",
            "title": "Depth"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "sender",
          "receiver",
          "depth"
        ],
        "title": "PeerProbeRequest"
      },
...
      "RunCreated": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "queued"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "task_id"
        ],
        "title": "RunCreated"
      },
      "RunView": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "brief": {
            "type": "string",
            "title": "Brief",
            "default": ""
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "phase": {
            "type": "string",
            "title": "Phase"
          },
          "model_mode": {
            "type": "string",
            "title": "Model Mode"
          },
          "behavior_version": {
            "type": "string",
            "title": "Behavior Version"
          },
          "event_count": {
            "type": "integer",
            "title": "Event Count",
            "default": 0
          },
          "last_cursor": {
            "type": "integer",
            "title": "Last Cursor",
            "default": 0
          },
          "completion_gate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completion Gate"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "run_id",
          "task_id",
          "title",
          "status",
          "phase",
          "model_mode",
          "behavior_version"
        ],
        "title": "RunView"
      },
...
      "SnapshotView": {
        "properties": {
          "run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunView"
              },
              {
                "type": "null"
              }
            ]
          },
          "agents": {
            "items": {
              "$ref": "#/components/schemas/AgentView"
            },
            "type": "array",
            "title": "Agents"
          },
          "assignments": {
            "items": {
              "$ref": "#/components/schemas/AssignmentView"
            },
            "type": "array",
            "title": "Assignments"
          },
          "contexts": {
            "items": {
              "$ref": "#/components/schemas/ContextView"
            },
            "type": "array",
            "title": "Contexts"
          },
          "capability_manifests": {
            "items": {
              "$ref": "#/components/schemas/CapabilityManifestView"
            },
            "type": "array",
            "title": "Capability Manifests"
          },
          "memories": {
            "items": {
              "$ref": "#/components/schemas/MemoryView"
            },
            "type": "array",
            "title": "Memories"
          },
          "evolutions": {
            "items": {
              "$ref": "#/components/schemas/EvolutionView"
            },
            "type": "array",
            "title": "Evolutions"
          },
          "dream_jobs": {
            "items": {
              "$ref": "#/components/schemas/DreamJobView"
            },
            "type": "array",
            "title": "Dream Jobs"
          },
          "runtime": {
            "$ref": "#/components/schemas/RuntimeView"
          }
        },
        "type": "object",
        "required": [
          "run",
          "agents",
          "assignments",
          "contexts",
          "capability_manifests",
          "memories",
          "evolutions",
          "dream_jobs",
          "runtime"
        ],
        "title": "SnapshotView"
      },
      "TaskRequest": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Title"
          },
          "brief": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Brief"
          },
          "model_mode": {
            "type": "string",
            "enum": [
              "scripted",
              "deepseek"
            ],
            "title": "Model Mode",
            "default": "scripted"
          },
          "execution_mode": {
            "type": "string",
            "enum": [
              "team",
              "single"
            ],
            "title": "Execution Mode",
            "default": "team"
          },
          "task_pack": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "resident-demo",
                  "repo-maintainer"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Pack"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "brief"
        ],
        "title": "TaskRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}